refactor: Merge types/docs from `bearer.pyi` into `crudclient/auth/bearer.py`
- Merged type hints and docstrings from `crudclient/auth/bearer.pyi` into `crudclient/auth/bearer.py`.

- Deleted `crudclient/auth/bearer.pyi`.
